improve record package
