+++ internal/record/filter_headers_test.go
@@ -1,4 +1,4 @@
-package actions
+package record
 
 import (
 	"reflect"
